Исправление бага HomeWorkdaoImpl

+++ src/main/java/ru/top/project/service/HomeWorkService.java
@@ -3,6 +3,7 @@ package ru.top.project.service;
 import ru.top.project.dao.HomeWorkDao;
 import ru.top.project.model.HomeWork;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class HomeWorkService {
@@ -22,11 +23,11 @@ public class HomeWorkService {
         homeWorkDao.updateHomeWork(homeWork);
     }
 
-    public void deleteHomeWork(int homeWorkId) {
+    public void deleteHomeWork(BigInteger homeWorkId) {
         homeWorkDao.deleteHomeWork(homeWorkId);
     }
 
-    public HomeWork getHomeWorkById(int homeWorkId) {
+    public HomeWork getHomeWorkById(BigInteger homeWorkId) {
         return homeWorkDao.getHomeWorkById(homeWorkId);
     }
